+++ lib/passport.js
@@ -54,12 +54,3 @@ passport.use('local.login', new LocalStrategy({
         }
     });
 }));
-
-module.exports.isAdmin = function() {
-    return isAdmin
-}
-
-module.exports.adminLogout = function() {
-    isAdmin = false
-    return
-}
