+++ models/song.js
@@ -15,8 +15,7 @@ var songSchema = new Schema({
     lyric: [[String]],
     source: Schema.Types.ObjectId,
     oriSong: String,
-    timeAdded: Date,
-    private: Boolean
+    timeAdded: Date
 })
 
 module.exports = mongoose.model('Song', songSchema)
