--- public/javascripts/languages.js
@@ -1,1 +1,0 @@
-var LANGUAGE_OPTIONS = ['english', 'mandarin', 'spanish', 'portuguese', 'bahasa']
