+++ routes/index.js
@@ -39,7 +39,9 @@ router.get('/', function(req, res, next) {
     var messages = req.flash();
     var langsExist;
     Song.find({
-            private: false
+            copyright: {
+                $ne: 'private'
+            }
         }, function(err, songs, count) {
             if (err) {
                 res.status(400).send('error getting song list ' + err)
@@ -242,7 +244,9 @@ router.get('/search', function(req, res, next) {
                             $options: 'si'
                         }
                     }],
-                    private: false
+                    copyright: {
+                        $ne: 'private'
+                    }
                 },
                 function(err, songs) {
                     done(null, songs)
@@ -302,7 +306,9 @@ router.get('/discover', function(req, res, next) {
     var langsExist;
     var messages = req.flash()
     Song.find({
-            private: false
+            copyright: {
+                $ne: 'private'
+            }
         }, function(err, songs, count) {
             if (err) {
                 res.status(400).send('error getting song list ' + err)
@@ -379,7 +385,9 @@ router.put('/discover', function(req, res, next) {
                             $options: 'si'
                         }
                     }],
-                    private: false
+                    copyright: {
+                        $ne: 'private'
+                    }
                 },
                 function(err, songs) {
                     done(err, songs)
@@ -389,7 +397,9 @@ router.put('/discover', function(req, res, next) {
                     source: {
                         $exists: false
                     },
-                    private: false
+                    copyright: {
+                        $ne: 'private'
+                    }
                 },
                 function(err, songs) {
                     done(err, songs)
@@ -408,7 +418,9 @@ router.put('/discover', function(req, res, next) {
                     temp.push(s)
                     Song.find({
                         source: s._id,
-                        private: false
+                        copyright: {
+                            $ne: 'private'
+                        }
                     }, (err, translations) => {
                         translations.forEach((t) => {
                             temp.push(t)
@@ -479,174 +491,6 @@ router.put('/discover', function(req, res, next) {
     async.waterfall([findOriginalSong, findTranslations, applyFilter, concatSongs, loadMore], finalize)
 
 })
-
-router.route('/song/:song_id')
-    .all(function(req, res, next) {
-        lang = req.query.lang || ''
-            // v = req.query.v || ''
-        song_id = req.params.song_id
-        song = {}
-        Song.findById(song_id, function(err, s) {
-            song = s;
-            next()
-        })
-    })
-    .get(function(req, res, next) {
-        Song.find({
-            $or: [{
-                $and: [{
-                    source: song.source
-                }, {
-                    source: {
-                        $exists: true
-                    }
-                }, {
-                    lang: {
-                        $ne: song.lang
-                    }
-                }]
-            }, {
-                _id: song.source
-            }, {
-                source: song._id
-            }]
-        }, function(err, translations) {
-            if (err) {
-                res.status(400).send('Error getting songs ' + err)
-            }
-
-            //rightTranslation is the song obj in the language that the user picks in the dropdown
-            var rightTranslation = translations.find((translation) => translation.lang === lang) || {}
-            var isTranslationExisted = !_.isEmpty(rightTranslation)
-            if (req.isAuthenticated()) {
-                Playlist.find({
-                    owner: req.user._id,
-                    song: {
-                        $exists: false
-                    }
-                }, function(err, playlists) {
-                    User.findOne({
-                        _id: req.user._id
-                    }, function(err, user) {
-                        console.log(user.library)
-                        res.render('song', {
-                            song: song,
-                            rightTranslation: rightTranslation,
-                            isTranslationExisted: isTranslationExisted,
-                            translations: translations,
-                            playlists: playlists,
-                            inLibrary: user.library
-                        })
-                    })
-
-                })
-            } else {
-                res.render('song', {
-                    song: song,
-                    rightTranslation: rightTranslation,
-                    isTranslationExisted: isTranslationExisted,
-                    translations: translations,
-                    playlists: [],
-                    inLibrary: []
-                })
-            }
-        })
-    })
-    //choosing translations
-    .put(function(req, res, next) {
-        var lang = req.body.lang
-        var leftColumnSongID = req.body.parentSongID
-        Song.findById(leftColumnSongID, function(err, song) {
-            if (err) return next(err)
-            if (!song.source) {
-                //then leftColumnSong is parent song in the Song Schema
-                //find one
-                Song.findOne({
-                    source: leftColumnSongID,
-                    lang: lang
-                }, function(err, translation) {
-                    if (err) return next(err)
-                    res.send({
-                        translation: translation
-                    })
-                })
-            } else {
-                Song.findOne({
-                    lang: lang,
-                    $or: [{
-                        source: song.source,
-                    }, {
-                        _id: song.source
-                    }]
-                })
-            }
-        })
-    })
-
-router.route('/song/:song_id/add-translation')
-    .all(function(req, res, next) {
-        song_id = req.params.song_id;
-        song = {}
-        Song.findById(song_id, function(err, s) {
-            song = s;
-            next();
-        })
-    })
-    .get(function(req, res, next) {
-        res.render('addTranslation', {
-            song: song
-        })
-    })
-    .post(function(req, res, next) {
-        req.checkBody('title', 'Title is empty').notEmpty()
-        req.checkBody('author', 'Author is empty').notEmpty()
-        req.checkBody('year', 'Year is empty').notEmpty()
-        var errors = req.validationErrors()
-        if (errors) {
-            res.send({
-                errorMessages: errors.map((error) => error.msg)
-            })
-        } else {
-            var data = req.body
-                //find for all cases, where song is parent song or child song
-            Song.findOne({
-                $or: [{
-                    source: song.source
-                }, {
-                    _id: song.source
-                }, {
-                    source: song._id
-                }],
-                lang: data.lang
-            }, function(err, translation) {
-                if (err) next(err)
-                if (translation || song.lang === data.lang) {
-                    res.send({
-                        errorMessages: ['Translation Exists']
-                    })
-                } else {
-                    var newSong = new Song({
-                        title: data.title,
-                        author: data.author,
-                        year: data.year,
-                        lang: data.lang,
-                        lyric: data.lyric,
-                        contributor: req.user.username,
-                        copyright: data.copyright,
-                        timeAdded: Date.now(),
-                        private: false,
-                        source: song._id
-                    })
-                    newSong.save(function(err) {
-                        if (err) next(err)
-                        res.send({
-                            url: '/song/' + song._id
-                        })
-                    })
-                }
-            })
-        }
-    })
 
 module.exports = router;
