--- views/addPrivateSong.jade
@@ -1,15 +1,0 @@
-extends layout
-
-block content
-    include topnav
-    .row
-        .col-md-6
-            h1(style='display:inline-block') Add a Private Song
-            button.btn.btn-default.pull-right(style='display:inline-block; margin-top: 25px' data-toggle='modal' data-target='#copypaste') Copy/Paste Lyric
-            #privateSong
-
-    script.
-        var song = !{JSON.stringify(song).replace(/<\//g, "<\\/")}
-    script(src='/dist/javascripts/privateSongForm.js')
-    script.
-        m.mount(document.getElementById('privateSong'), privateSongForm(song, 'add', false).privateSongFormComponent)
