--- views/addPrivateSongTranslation.jade
@@ -1,49 +1,0 @@
-extends layout
-
-block content
-    include topnav
-    - var langCapitalized = song.lang.charAt(0).toUpperCase() + song.lang.slice(1)
-    //- if messages[0]
-    //-     .alert.alert-danger
-    //-         each msg in messages
-    //-             p #{msg.msg}
-        
-    form(role='form', method="post", action="")
-        .row
-            .col-md-6
-                h1 Original Song
-            .col-md-6
-                h1 Translation
-                if hasErrors
-                    .alert.alert-danger 
-                        each msg in messages
-                            p #{msg}      
-        .row
-            .col-md-6
-                .form-group
-                    label(for="lang") Language
-                    input#language.form-control(name="lang", value=langCapitalized, readonly)         
-                .form-group
-                    label(for="title") Title
-                    input#title.form-control(name="title", type="text", value=song.title, readonly)  
-            .col-md-6
-                .form-group
-                    label(for="translationLang") Language
-                    select#language.form-control(name="translationLang")
-                        option(value="english") English
-                        option(value="mandarin") Mandarin
-                        option(value="spanish") Spanish
-                        option(value="portuguese") Portuguese
-                .form-group
-                    label(for="translationTitle") Title
-                    input#title.form-control(name="translationTitle", type="text")
-        .row
-            .col-md-6
-                .form-group
-                    label(for="lyric") Lyric
-                    textarea#lyric.form-control(name="lyric", rows='20' readonly) #{stringLyric}
-            .col-md-6
-                .form-group
-                    label(for="translationLyric") Lyric
-                    textarea#lyric.form-control(name="translationLyric", rows='20')
-                button.btn.btn-primary(type="submit") Submit Translation
