+++ views/addTranslation.jade
@@ -14,6 +14,7 @@ block content
 
     script.
         var song = !{JSON.stringify(song).replace(/<\//g, "<\\/")}
+        var availableLanguages = !{JSON.stringify(availableLanguages).replace(/<\//g, "<\\/")}
     script(src='/dist/javascripts/songForm.js')
     script.
         var translationObject = {title: '', author: '', year: '', lang: 'english', copyright: 'CC0', lyric: []}
