+++ views/library.jade
@@ -8,7 +8,7 @@ block content
         .alert.alert-danger
             p #{messages.error}
     br
-    a.btn.btn-default(href='privatesong/add') Add Private Song
+    a.btn.btn-default(href='/songlist-db/add') Add Song
     #library
 
     script.
